UI: Refactor window input into a separate module
Create a zygrader/ui/input.py file to handle the input thread
separate from the window. Yay for code quality!

At the moment, resizing, the VIM header indicator, and the christmas
theme alternating events are broken. This will be fixed more with more
refactoring of the event loop.
