+++ zygrader/grader.py
@@ -44,7 +44,8 @@ def fill_student_list(lab, students):
 def update_student_list(window: ui.Window, student_list: ui.components.FilteredList):
     """Update the student list when the locks or flags change"""
     student_list.refresh()
-    window.push_refresh_event()
+    input = ui.get_input()
+    input.push_refresh_event()
 
 
 def get_submission(lab, student, use_locks=True):
