+++ zygrader/ui/__init__.py
@@ -1,4 +1,5 @@
-from .window import Window, WinContext, Event, GO_BACK
+from .window import Window, WinContext
+from .input import Event, GO_BACK, Input
 from . import components
 from . import window
 
@@ -6,3 +7,6 @@ from . import window
 def get_window() -> Window:
     """Utility function to make getting the Window simpler in code"""
     return Window.get_window()
+
+def get_input() -> Input:
+    return get_window().input
